Introduce PerspectiveService as part of system_server
This moves all the perspective logic from PerspectiveManager to
a central service that runs in system_server. This is the correct
way to do things and is much more efficient as there is only one
service no matter the number of clients (previously we would create
a new PerspectiveManager for each client that ran all the logic).

PerspectiveService relies on a state transition model for perspectives
to keep everyone on the same page. A perspective can be in either the
terminal states of STOPPED or RUNNING, or be in a transitionary state
like STARTING or STOPPING. Clients can choose to be kept informed of these
changes in real-time.

(cherry picked from commit 2a7731403cc7eb68f9b751215b7828572ff32262,
added updated vendor include directory and squashed static version of
IInterface::asBinder())
Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>