Update desktop QS to new perspective API but remove tile for now
This updates MDesktopTile to use the new APIs introduced from the
PerspectiveService system_server change.

This also removes MDesktopTile from the QS panel for now since it
is redundant with the Maru Desktop notification and link to the
Desktop settings fragment. Still keeping the code here if users would
like a shortcut to quickly control the desktop though.