maru: Grant storage perms to Browser by default
This fixes the Browser crash when attempting to download a file off of a
fresh system. This was due to the Browser not having storage access by
default, and (probably) a bug in the Browser code that does not check to
make sure it has storage perms when downloading a file.

Note that this should only affect the default browser app shipped on
/system. There is a separate function for granting default permissions
to a new browser app that the user sets to be the default browser, and
that is left untouched (only grants location perms).

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>