Add desktop settings category
This can be used to implement a dynamic setting under the main Settings
> Desktop dashboard.

We currently do not make use of this, but will likely in the future when
Maru adds the capability to select different desktop rootfs images. This
will allow the logic to live in a separate activity and by dynamically
loaded into Settings under the main desktop settings dashboard.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>