Enable dynamic input switching
When desktop is interactive, we will disable external input devices, and
enable those devices when desktop is non-interactive.

Default, the Android don't show IME when hardware
keyboard is connected. But when desktop is interactive,
the hardware keyboard event is directed to desktop,
so we need showing IME when desktop is interactive
for better user experience.

If device list changed, we will trigger another desktop state changed
event to ensure the state of external input devices is correctly.

Note: this CL is squashed from three commits about dynamic input
switching from maru-0.6, and use new API of InputManager to simply the
logic.

Signed-off-by: utzcoz <utzcoz@outlook.com>