Disable unhandled key when desktop is interactive
If we press ESC more than one time when desktop is interactive,
the system will switch to launcher in Android, which is caused
by unhandled key. So we need to disable unhandled key when
desktop is interactive to do more key interception.

Also it will help intercept defined shortcut, such as
ALT + TAB for switching app.

Signed-off-by: utzcoz <utzcoz@outlook.com>